[feat]: useRemainingFunding 훅 구현(#203)
- 일정 금액을 입력받아 잔여금액을 관리하는 훅 구현
